+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Billing: Added optional fields description and short description to products grid in catalogue
 - Email: Fixed &euro; shown in subject instead of €
 - Core: Link/changes could lead to double join of core_search s
+- Core: Improved list style in HTML editor and SHIFT + ENTER in lists inserts line break
+  instead of new list item.
 
 26-02-2021 6.4.224
 - Multi instance: Fixed delete issue when temp dir was in use